Update to distributed SPICE model simulation
We have refactored the netlist generation code to use model objects that
generate a netlist. The netlist is solved using PySpice. This removes
the need for code in Solcore to kick off a simulation and process the
result using text parsing. I have not looked too closely, but parser
code is usually quite brittle; changes to the net list can break it.
Handing this over to PySpice improves the situation.

A result object is passed around to helper functions to generate useful
information such as the IV curve, maximum power point, voltages of the
different layers, and electroluminescence prediction.

The distributed model closely aligns with the methodology of Steiner
et al. DOI:10.1002/pip.989, except for the absence of the perimeter
recombination diode. This model lacks shunt resistances, but adding them
could broaden its application beyond concentrator solar cells.

This code is standalone; it uses basic Python types to define the solar
cell's junction rather than using Solcore's Junction class. Future
changes may want to alter this, but I'm uncertain about the value that
Junction offers; it seems to resemble an OrderedDict. Perhaps this would
unlock useful integration with other Solcore features.

A nice feature of the code is a class for generating grid metalization
patterns: GridPattern. This offers a general interface for subclassing,
enabling the creation of various grid types. The user needs to implement
a function called `draw()` and use the pixie-python drawing API to
render the desired grid pattern. A HGridPattern subclass has been
implemented here. This class offers some useful functionality, such as
being able to save the grid pattern to an image file, which uses pillow
(PIL).

We have added a sparse example script. We have added a much more
detailed Python Notebook file that guides the user through the entire
process, from basic simulation to computing a concentration vs.
efficiency plot.